floattt

+++ Fulfil.Dispense/src/commands/pre_side_dispense/pre_side_dispense_response.cpp
@@ -35,7 +35,7 @@ void PreSideDispenseResponse::encode_payload() {
 
 PreSideDispenseResponse::PreSideDispenseResponse(std::shared_ptr<std::string> request_id,
                                                  std::shared_ptr<std::string> primary_key_id,
-                                                 std::shared_ptr<std::vector<std::vector<int> > > occupancy_map,
+                                                 std::shared_ptr<std::vector<std::vector<float> > > occupancy_map,
                                                  SideDispenseErrorCodes success_code,
                                                  std::string error_description) : request_id(request_id),
     primary_key_id(primary_key_id),
@@ -62,7 +62,7 @@ std::shared_ptr<std::string> PreSideDispenseResponse::dispense_payload() {
     return this->payload;
 }
 
-std::string grid_map_to_string(std::shared_ptr<std::vector<std::vector<int> > > map) {
+std::string grid_map_to_string(std::shared_ptr<std::vector<std::vector<float> > > map) {
     if (map == nullptr) return "nullptr";
     std::string output = "";
     std::string val;
